requirements.txt
# ── Web scraping ──────────────────────────────────────────────────────────────
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0

# ── Text to Speech ────────────────────────────────────────────────────────────
gTTS==2.5.1

# ── Image processing ──────────────────────────────────────────────────────────
Pillow==10.3.0

# ── YouTube API ───────────────────────────────────────────────────────────────
google-api-python-client==2.118.0
google-auth==2.28.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
